Linked Lists
The first SinglyLinkedList implementation does not use a “size”
variable, and calls the size() method when it needs it. This is not
efficient, as the list is traversed more than necessary.
The second SinglyLinkedList implementation still doesn’t use a “size”
variable, but instead includes the size verification within the methods
that need it. This prevents unnecessary traversal of the list.
The third SinglyLinkedList implementation uses an instance variable
“size” instead.
The “tail” implementation provide a fast implementation of addLast.
The “doublyLinked” implementation adds a fast implementation of
removeLast.
The “DummyNode” implementation doesn’t improve speed, but simplifies
the code.